requirements.txt
django
pytest-django
pytest-factoryboy
pytest-faker
selenium